chore: rename sample app package and use public repos
