+++ .editorconfig
@@ -1,7 +1,7 @@
 root = true
 
 [*.{kt,kts}]
-disabled_rules = package-name,filename
+disabled_rules = filename
 max_line_length = off
 insert_final_newline = true
 ij_kotlin_name_count_to_use_star_import = 999
