+++ sample/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.kroger.sample_app">
+    package="com.kroger.sample">
 
     <application
-        android:name="com.kroger.sample_app.SampleApplication"
+        android:name="com.kroger.telemetry.sample.SampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Telemetry">
         <activity
-            android:name="com.kroger.sample_app.MainActivity"
+            android:name="com.kroger.telemetry.sample.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
