+++ sample/src/main/java/com/kroger/telemetry/sample/BarRelay.kt
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.kroger.sample_app
+package com.kroger.telemetry.sample
 
 import android.util.Log
 import com.kroger.telemetry.TypedRelay
