+++ sample/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kroger.sample_app.MainActivity">
+    tools:context="com.kroger.sample.MainActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/event_button"
